Add new methods to handle AWS responses in a standard way
-resource_fail
This method should be used when AWS API returns multiple resources for the provided criteria.
it { should exist } test will fail and raise an error.

-empty_response_warn
This method should be used when the AWS API returns an empty response, e.g. `[]`.
it { should_not exist } test will pass.

-tags
This will ensure creating an empty hash value if the API does not return `tags` property even though it is supported. E.g. `subnets`.

-name
Expose the value of the `Name` tag if applied to the resource.

Signed-off-by: Omer Demirok <odemirok@chef.io>