Works around added init data
The addition of the last line in the init method seems to be breaking
the resource pack becaue it doesn't account for the nil? case. Added a
protection for this.

Review of why and for what function this was added 2 years ago may be
useful but doesn't seem to be core to the function of the resources
overall.

Fixes #977

Signed-off-by: Aaron Lippold <lippold@gmail.com>